+++ src/main/java/de/unistuttgart/iste/gits/user_service/persistence/entity/CourseMembershipPk.java
@@ -1,9 +1,6 @@
-package de.unistuttgart.iste.gits.user_service.persistence.dao;
+package de.unistuttgart.iste.gits.user_service.persistence.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.io.Serializable;
 import java.util.UUID;
