+++ src/main/java/de/unistuttgart/iste/gits/user_service/service/MembershipService.java
@@ -5,8 +5,8 @@ import de.unistuttgart.iste.gits.common.event.CrudOperation;
 import de.unistuttgart.iste.gits.generated.dto.CourseMembership;
 import de.unistuttgart.iste.gits.generated.dto.CourseMembershipInput;
 import de.unistuttgart.iste.gits.user_service.mapper.MembershipMapper;
-import de.unistuttgart.iste.gits.user_service.persistence.dao.CourseMembershipEntity;
-import de.unistuttgart.iste.gits.user_service.persistence.dao.CourseMembershipPk;
+import de.unistuttgart.iste.gits.user_service.persistence.entity.CourseMembershipEntity;
+import de.unistuttgart.iste.gits.user_service.persistence.entity.CourseMembershipPk;
 import de.unistuttgart.iste.gits.user_service.persistence.repository.CourseMembershipRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
