+++ src/main/resources/graphql/service/courseMembership.graphqls
