+++ src/main/resources/graphql/service/query.graphqls
