+++ src/main/resources/graphql/service/userInfo.graphqls
